requirements.txt
Django==3.0.6
psycopg2==2.8.5
Pillow==7.1.2
django-resized==0.3.11
django-summernote==0.8.11.6
django-jalali-date==0.3.1
gunicorn==20.0.4
django-admin-interface==0.12.2
requests==2.24.0